FIX: hack to make tracker search work with 3.6.1
Ubuntu 12.10 ships with gnome-shell 3.6.1. This hack makes the
extension work on 12.10.

Technical Debt: the uuid of the extension has been changed to match
the folder that is defined by the project. This doesn't look like
other extensions that are commonly available so fixing the folder
name and then the uuid in the metadata.json remains an open task.